data-in/requirements.txt
pydantic>=2.4
PyYAML>=6.0
httpx>=0.24
pdfplumber>=0.10
pdf2image>=1.16
pytesseract>=0.3
Pillow>=10.0
PyMuPDF>=1.23
